requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
configparser==5.0.2
cryptography==3.4.7
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
idna==2.10
jsonfield==3.1.0
lxml==4.6.3
mailer==0.8.1
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.20.3
pathspec==0.8.1
Pillow==8.2.0
psycopg2
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
python-pptx==0.6.19
pytz==2021.1
regex==2021.7.6
reportlab==3.5.67
requests==2.25.1
scipy==1.7.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
XlsxWriter==1.4.3
